Move `AverageError` into `mlp` module
Never mind it's an expert overfitting tool.
